requirements.txt
streamlit==1.25.0
pytesseract==0.3.9
Pillow==10.0.0
